admin user  edit function added

+++ app/Http/Livewire/Backend/AccessControl/RoleTable.php
@@ -9,7 +9,7 @@ use Spatie\Permission\Models\Role;
 class RoleTable extends Component
 {
   use WithPagination;
-  protected $paginationTheme = 'bootstrap'; 
+  protected $paginationTheme = 'bootstrap';
   protected $listeners = ['remove'];
 
 
@@ -19,16 +19,16 @@ class RoleTable extends Component
    * @return response()
    */
 
-    public function alertConfirm($roleId)
-    {
+  public function alertConfirm($roleId)
+  {
 
-        $this->dispatchBrowserEvent('swal:confirm', [
-                'type' => 'warning',  
-                'message' => 'Are you sure?', 
-                'text' => 'If deleted, you will not be able to recover the data',
-                'roleId' => $roleId
-            ]);
-    }
+    $this->dispatchBrowserEvent('swal:confirm', [
+      'type' => 'warning',
+      'message' => 'Are you sure?',
+      'text' => 'If deleted, you will not be able to recover the data',
+      'roleId' => $roleId
+    ]);
+  }
 
   /**
    * Write code on Method
@@ -36,22 +36,23 @@ class RoleTable extends Component
    * @return response()
    */
 
-    public function remove($roleId)
-    {
+  public function remove($roleId)
+  {
+    if ($roleId != 1) {
       $distroyRole = Role::findOrFail($roleId);
       $distroyRole->delete();
-        /* Write Delete Logic */
-        $this->dispatchBrowserEvent('swal:modal', [
-                'type' => 'success',  
-                'message' => 'Role Deleted Successfully!', 
-                'text' => 'regret wont be enough ( ͡❛ ₃ ͡❛)'
-            ]);
-
+      /* Write Delete Logic */
+      $this->dispatchBrowserEvent('swal:modal', [
+        'type' => 'success',
+        'message' => 'Role Deleted Successfully!',
+        'text' => 'regret wont be enough ( ͡❛ ₃ ͡❛)'
+      ]);
     }
+  }
+  
 
-
-    public function render()
-    {
-        return view('livewire.backend.access-control.role-table', ['roles' => Role::with('permissions')->paginate(10)]);
-    }
+  public function render()
+  {
+    return view('livewire.backend.access-control.role-table', ['roles' => Role::with('permissions')->paginate(10)]);
+  }
 }
